add tests for fornax-nb
